+++ extension/src/PermissionSet/XmlPermissionSet.ts
@@ -1,4 +1,6 @@
 import * as xmldom from "@xmldom/xmldom";
+import { ALObjectType } from "../ALObject/Enums";
+import { alObjectTypeMap, alObjectTypeNumberMap } from "../ALObject/Maps";
 
 export class XmlPermissionSets {
   permissionSets: XmlPermissionSet[] = [];
@@ -27,35 +29,51 @@ export class XmlPermissionSets {
         filePath: filePath,
         suggestedNewName: prefix + roleId,
       };
-      const permissionsNodeList = permissionSetNode.getElementsByTagName(
+      let permissionsNodeList = permissionSetNode.getElementsByTagName(
         "Permission"
       );
+      if (permissionsNodeList.length === 0) {
+        permissionsNodeList = permissionSetNode.getElementsByTagName(
+          "TenantPermission"
+        );
+      }
       for (let i = 0; i < permissionsNodeList.length; i++) {
         const permissionsNode = permissionsNodeList[i];
-        const objectType = permissionsNode.getElementsByTagName("ObjectType")[0]
-          .childNodes[0].nodeValue;
-
-        const objectId = permissionsNode.getElementsByTagName("ObjectID")[0]
-          .childNodes[0].nodeValue;
-        const readPermission =
-          permissionsNode.getElementsByTagName("ReadPermission")[0]
-            .childNodes[0].nodeValue ?? "";
-        const insertPermission =
-          permissionsNode.getElementsByTagName("InsertPermission")[0]
-            .childNodes[0].nodeValue ?? "";
-        const modifyPermission =
-          permissionsNode.getElementsByTagName("ModifyPermission")[0]
-            .childNodes[0].nodeValue ?? "";
-        const deletePermission =
-          permissionsNode.getElementsByTagName("DeletePermission")[0]
-            .childNodes[0].nodeValue ?? "";
-        const executePermission =
-          permissionsNode.getElementsByTagName("ExecutePermission")[0]
-            .childNodes[0].nodeValue ?? "";
+        const objectType = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "ObjectType"
+        );
+        const objectId = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "ObjectID"
+        );
+        const readPermission = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "ReadPermission"
+        );
+        const insertPermission = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "InsertPermission"
+        );
+        const modifyPermission = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "ModifyPermission"
+        );
+        const deletePermission = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "DeletePermission"
+        );
+        const executePermission = XmlPermissionSets.getPermissionValue(
+          permissionsNode,
+          "ExecutePermission"
+        );
+        const securityFilter = permissionsNode.getElementsByTagName(
+          "SecurityFilter"
+        );
         if (
-          permissionsNode
-            .getElementsByTagName("SecurityFilter")[0]
-            .hasChildNodes()
+          securityFilter &&
+          securityFilter[0] &&
+          securityFilter[0].hasChildNodes()
         ) {
           throw new Error(
             `PermissionSet ${permissionSet.roleID} har defined some SecurityFilter, which is unsupported by this function`
@@ -72,18 +90,68 @@ export class XmlPermissionSets {
           );
         }
         const permission: XmlPermission = {
-          objectType: objectType,
+          objectType: this.getALObjectType(objectType),
           objectID: Number.parseInt(objectId),
-          readPermission: readPermission,
-          insertPermission: insertPermission,
-          modifyPermission: modifyPermission,
-          deletePermission: deletePermission,
-          executePermission: executePermission,
+          readPermission: XmlPermissionSets.stringToPermission(readPermission),
+          insertPermission: XmlPermissionSets.stringToPermission(
+            insertPermission
+          ),
+          modifyPermission: XmlPermissionSets.stringToPermission(
+            modifyPermission
+          ),
+          deletePermission: XmlPermissionSets.stringToPermission(
+            deletePermission
+          ),
+          executePermission: XmlPermissionSets.stringToPermission(
+            executePermission
+          ),
         };
         permissionSet.permissions.push(permission);
       }
       this.permissionSets.push(permissionSet);
     }
+  }
+
+  static getPermissionValue(
+    permissionsNode: Element,
+    permission: string
+  ): string {
+    const node = permissionsNode.getElementsByTagName(permission);
+    if (!node || !node[0] || !node[0].childNodes || !node[0].childNodes[0]) {
+      return "";
+    }
+    return node[0].childNodes[0].nodeValue ?? "";
+  }
+
+  static stringToPermission(permissionString: string): Permission {
+    switch (permissionString.toLocaleLowerCase()) {
+      case "0":
+      case "":
+        return Permission.none;
+      case "1":
+      case "yes":
+        return Permission.yes;
+      case "2":
+      case "indirect":
+        return Permission.indirect;
+      default:
+        throw new Error(`Unexpected permission value: ${permissionString}`);
+    }
+  }
+
+  private getALObjectType(objectTypeText: string): ALObjectType {
+    const alObjectType = alObjectTypeMap.get(
+      objectTypeText.replace(" ", "").toLowerCase()
+    );
+    if (alObjectType) {
+      return alObjectType;
+    }
+    const objectTypeNumber: number = Number.parseInt(objectTypeText);
+    const objectType = alObjectTypeNumberMap.get(objectTypeNumber);
+    if (!objectType) {
+      throw new Error(`No object type found for "${objectTypeText}"`);
+    }
+    return objectType;
   }
 }
 
@@ -96,11 +164,17 @@ export interface XmlPermissionSet {
 }
 
 export interface XmlPermission {
-  objectType: string;
+  objectType: ALObjectType;
   objectID: number;
-  readPermission: string;
-  insertPermission: string;
-  modifyPermission: string;
-  deletePermission: string;
-  executePermission: string;
+  readPermission: Permission;
+  insertPermission: Permission;
+  modifyPermission: Permission;
+  deletePermission: Permission;
+  executePermission: Permission;
+}
+
+export enum Permission {
+  none = 0,
+  yes = 1,
+  indirect = 2,
 }