+++ extension/src/test/PermissionSet.test.ts
@@ -4,8 +4,9 @@ import * as assert from "assert";
 import * as FileFunctions from "../FileFunctions";
 import * as SettingsLoader from "../Settings/SettingsLoader";
 import * as PermissionSetFunctions from "../PermissionSet/PermissionSetFunctions";
-
 import { getPermissionSetFiles } from "../WorkspaceFunctions";
+import { ALObjectType } from "../ALObject/Enums";
+import { Permission } from "../PermissionSet/XmlPermissionSet";
 
 const testOrgFiles = path.resolve(
   __dirname,
@@ -25,14 +26,14 @@ suite("PermissionSet", function () {
 
   test("Parse PermissionSet XML Files", async function () {
     const filePaths = getPermissionSetFiles(testFilesPath);
-    assert.strictEqual(filePaths.length, 2, "Unexpected number of files");
+    assert.strictEqual(filePaths.length, 3, "Unexpected number of files");
     const xmlPermissionSets = await PermissionSetFunctions.getXmlPermissionSets(
       filePaths,
       ""
     );
     assert.strictEqual(
       xmlPermissionSets.length,
-      2,
+      3,
       "Unexpected number of permission sets"
     );
     assert.strictEqual(
@@ -47,9 +48,101 @@ suite("PermissionSet", function () {
     );
     assert.strictEqual(
       xmlPermissionSets[0].permissions.length,
-      22,
-      "Unexpected number of permissions"
+      26,
+      "Unexpected number of permissions [0]"
     );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[0],
+      {
+        objectType: ALObjectType.tableData,
+        objectID: 50002,
+        readPermission: Permission.yes,
+        insertPermission: Permission.yes,
+        modifyPermission: Permission.yes,
+        deletePermission: Permission.yes,
+        executePermission: Permission.none,
+      },
+      "Unexpected permission[0]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[1],
+      {
+        objectType: ALObjectType.table,
+        objectID: 50002,
+        readPermission: Permission.none,
+        insertPermission: Permission.none,
+        modifyPermission: Permission.none,
+        deletePermission: Permission.indirect,
+        executePermission: Permission.yes,
+      },
+      "Unexpected permission[1]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[2],
+      {
+        objectType: ALObjectType.tableData,
+        objectID: 50003,
+        readPermission: Permission.yes,
+        insertPermission: Permission.indirect,
+        modifyPermission: Permission.yes,
+        deletePermission: Permission.yes,
+        executePermission: Permission.none,
+      },
+      "Unexpected permission[2]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[3],
+      {
+        objectType: ALObjectType.system,
+        objectID: 3510,
+        readPermission: Permission.none,
+        insertPermission: Permission.none,
+        modifyPermission: Permission.none,
+        deletePermission: Permission.none,
+        executePermission: Permission.yes,
+      },
+      "Unexpected permission[3]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[4],
+      {
+        objectType: ALObjectType.query,
+        objectID: 50001,
+        readPermission: Permission.none,
+        insertPermission: Permission.none,
+        modifyPermission: Permission.none,
+        deletePermission: Permission.none,
+        executePermission: Permission.yes,
+      },
+      "Unexpected permission[4]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[5],
+      {
+        objectType: ALObjectType.page,
+        objectID: 50007,
+        readPermission: Permission.none,
+        insertPermission: Permission.none,
+        modifyPermission: Permission.none,
+        deletePermission: Permission.none,
+        executePermission: Permission.yes,
+      },
+      "Unexpected permission[5]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[0].permissions[6],
+      {
+        objectType: ALObjectType.codeunit,
+        objectID: 50002,
+        readPermission: Permission.none,
+        insertPermission: Permission.none,
+        modifyPermission: Permission.none,
+        deletePermission: Permission.none,
+        executePermission: Permission.yes,
+      },
+      "Unexpected permission[6]"
+    );
+
     assert.strictEqual(
       xmlPermissionSets[1].roleID,
       "AL-2",
@@ -57,8 +150,37 @@ suite("PermissionSet", function () {
     );
     assert.strictEqual(
       xmlPermissionSets[1].roleName,
-      "Al 2",
+      "",
       "Unexpected roleName 1"
+    );
+
+    assert.strictEqual(
+      xmlPermissionSets[2].roleID,
+      "AL-TENANT",
+      "Unexpected roleId 0"
+    );
+    assert.strictEqual(
+      xmlPermissionSets[2].roleName,
+      "TenantPermissions",
+      "Unexpected roleName 0"
+    );
+    assert.strictEqual(
+      xmlPermissionSets[2].permissions.length,
+      5,
+      "Unexpected number of permissions [2]"
+    );
+    assert.deepStrictEqual(
+      xmlPermissionSets[2].permissions[0],
+      {
+        objectType: ALObjectType.tableData,
+        objectID: 50003,
+        readPermission: Permission.yes,
+        insertPermission: Permission.indirect,
+        modifyPermission: Permission.yes,
+        deletePermission: Permission.yes,
+        executePermission: Permission.none,
+      },
+      "Unexpected permission[2][0]"
     );
   });
 
@@ -71,6 +193,7 @@ suite("PermissionSet", function () {
       filePaths,
       prefix
     );
+    xmlPermissionSets[1].roleName = "A Name";
     PermissionSetFunctions.validateData(xmlPermissionSets);
     await PermissionSetFunctions.startConversion(
       prefix,